+++ src/cli/init_cmd.ml
@@ -1,7 +1,7 @@
 open Cmdliner
-open Core
+open Common
 open Console
-open Files
+open Module
 
 let with_file : string -> (out_channel -> unit) -> unit = fun file fn ->
   let oc = open_out file in fn oc; close_out oc