+++ src/handle/rewrite.ml
@@ -3,9 +3,11 @@
 open! Lplib
 
 open Timed
+open Common
 open Pos
-open Terms
-open Basics
+open Core
+open Term
+open LibTerm
 open Console
 open Print
 open Proof
@@ -58,18 +60,6 @@ let eq : ctxt -> term -> term -> bool = fun ctx a b -> a == b ||
     end
   in
   try eq [(a,b)]; true with Not_equal -> false
-
-(** Rewrite patterns as in Coq/SSReflect. See "A Small Scale
-    Reflection Extension for the Coq system", by Georges Gonthier,
-    Assia Mahboubi and Enrico Tassi, INRIA Research Report 6455, 2016,
-    @see <http://hal.inria.fr/inria-00258384>, section 8, p. 48. *)
-type rw_patt =
-  | RW_Term           of term
-  | RW_InTerm         of term
-  | RW_InIdInTerm     of (term, term) Bindlib.binder
-  | RW_IdInTerm       of (term, term) Bindlib.binder
-  | RW_TermInIdInTerm of term * (term, term) Bindlib.binder
-  | RW_TermAsIdInTerm of term * (term, term) Bindlib.binder
 
 (** Equality configuration. *)
 type eq_config =
@@ -173,11 +163,11 @@ let _ =
 let get_eq_data : popt -> eq_config -> term -> term * term * term =
   fun pos cfg t ->
   let t = Eval.whnf [] t in
-  match Basics.get_args t with
+  match LibTerm.get_args t with
   | (p, [u]) when is_symb cfg.symb_P p ->
       begin
         let u = Eval.whnf [] u in
-        match Basics.get_args u with
+        match LibTerm.get_args u with
         | (eq, [a;l;r]) when is_symb cfg.symb_eq eq -> (a, l, r)
         | _ ->
            fatal pos "Expected an equality type, found [%a]." pp_term t
@@ -346,7 +336,7 @@ let rewrite : Sig_state.t -> popt -> proof_state -> bool -> rw_patt option
 
   (* Extract the term from the goal type (get “t” from “P t”). *)
   let g_term =
-    match Basics.get_args g_type with
+    match LibTerm.get_args g_type with
     | (p, [t]) when is_symb cfg.symb_P p -> t
     | _                                        ->
         fatal pos "Goal type [%a] is not of the form “P t”." pp_term g_type