+++ src/handle/tactic.ml
@@ -2,10 +2,13 @@
 
 open! Lplib
 
+open Common
 open Console
-open Terms
 open Pos
+open Parsing
 open Syntax
+open Core
+open Term
 open Proof
 open Print
 open Timed
@@ -44,7 +47,7 @@ let tac_refine : popt -> proof_state -> term -> proof_state =
   | Typ gt::gs ->
       if !log_enabled then
         log_tact "refine %a with %a" pp_meta gt.goal_meta pp_term t;
-      if Basics.occurs gt.goal_meta t then fatal pos "Circular refinement.";
+      if LibTerm.occurs gt.goal_meta t then fatal pos "Circular refinement.";
       (* Check that [t] is well-typed. *)
       let gs_typ, gs_unif = List.partition is_typ gs in
       let to_solve = List.map get_constr gs_unif in
@@ -58,16 +61,16 @@ let tac_refine : popt -> proof_state -> term -> proof_state =
             (Bindlib.unbox (Bindlib.bind_mvar (Env.vars gt.goal_hyps)
                               (lift t)));
           (* Convert the metas of [t] not in [gs] into new goals. *)
-          let gs_typ = add_goals_of_metas (Basics.get_metas true t) gs_typ in
+          let gs_typ = add_goals_of_metas (LibTerm.get_metas true t) gs_typ in
           let proof_goals = List.rev_map (fun c -> Unif c) cs @ gs_typ in
           tac_solve pos {ps with proof_goals}
 
-(** [handle_tactic ss e ps tac] applies tactic [tac] in the proof state
+(** [handle ss e ps tac] applies tactic [tac] in the proof state
    [ps] and returns the new proof state. This function fails gracefully in
    case of error. *)
-let handle_tactic :
-  Sig_state.t -> Terms.expo -> proof_state -> p_tactic ->
-    proof_state * Queries.result =
+let handle :
+  Sig_state.t -> Tags.expo -> proof_state -> p_tactic ->
+    proof_state * Query.result =
   fun ss e ps tac ->
   if !log_enabled then
     begin
@@ -77,7 +80,7 @@ let handle_tactic :
       log_tact "%a" Pretty.tactic tac
     end;
   match tac.elt with
-  | P_tac_query(q) -> ps, Queries.handle_query ss (Some ps) q
+  | P_tac_query(q) -> ps, Query.handle ss (Some ps) q
   | P_tac_solve -> tac_solve tac.pos ps, None
   | P_tac_focus(i) ->
       (try {ps with proof_goals = List.swap i ps.proof_goals}, None
@@ -96,7 +99,7 @@ let handle_tactic :
       let n =
         match Infer.infer_noexn [] (Env.to_ctxt env) t with
         | None -> fatal tac.pos "[%a] is not typable." pp_term t
-        | Some (a, _) -> Basics.count_products a
+        | Some (a, _) -> LibTerm.count_products a
       in
       (*FIXME: this does not take into account implicit arguments. *)
       let t = if n <= 0 then t
@@ -124,9 +127,9 @@ let handle_tactic :
             None)
   | P_tac_fail -> fatal tac.pos "Call to tactic \"fail\""
 
-let handle_tactic :
-  Sig_state.t -> Terms.expo -> proof_state -> p_tactic ->
-    proof_state * Queries.result =
+let handle :
+  Sig_state.t -> Tags.expo -> proof_state -> p_tactic ->
+    proof_state * Query.result =
   fun ss exp ps tac ->
-  try handle_tactic ss exp ps tac
+  try handle ss exp ps tac
   with Fatal(_,_) as e -> out 1 "%a" Proof.pp_goals ps; raise e