+++ src/tool/xtc.ml
@@ -8,7 +8,8 @@ open Lplib.Base
 open Lplib.Extra
 
 open Timed
-open Terms
+open Core
+open Term
 
 (** [print_sym oc s] outputs the fully qualified name of [s] to [oc]. Modules
     are separated with ["."]. *)
@@ -24,7 +25,7 @@ type symb_status = Object_level | Basic_type | Type_cstr
 let status : sym -> symb_status = fun s ->
   (* the argument [b] of [is_arrow_kind] is a boolean saying if we have
      already gone under a product *)
-  let rec is_arrow_kind : Terms.term -> bool -> symb_status = fun t b ->
+  let rec is_arrow_kind : Term.term -> bool -> symb_status = fun t b ->
     match t with
     | Prod(_,b) -> is_arrow_kind (snd (Bindlib.unbind b)) true
     | Type      -> if b then Type_cstr else Basic_type
@@ -100,26 +101,26 @@ and print_type : int -> string -> term pp = fun i s oc t ->
 let print_rule : Format.formatter -> int -> sym -> rule -> unit =
   fun oc i s r ->
   (* Print the rewriting rule. *)
-  let lhs = Basics.add_args (Symb s) r.lhs in
+  let lhs = LibTerm.add_args (Symb s) r.lhs in
   Format.fprintf oc "<rule>@.<lhs>@.%a</lhs>@." (print_term i s.sym_name) lhs;
-  let rhs = Basics.term_of_rhs r in
+  let rhs = LibTerm.term_of_rhs r in
   Format.fprintf oc "<rhs>@.%a</rhs>@.</rule>@." (print_term i s.sym_name) rhs
 
 (** [print_tl_rule] is identical to [print_rule] but for type-level rule  *)
 let print_tl_rule : Format.formatter -> int -> sym -> rule -> unit =
   fun oc i s r ->
   (* Print the type level rewriting rule. *)
-  let lhs = Basics.add_args (Symb s) r.lhs in
+  let lhs = LibTerm.add_args (Symb s) r.lhs in
   Format.fprintf oc "<typeLevelRule>@.<TLlhs>@.%a</TLlhs>@."
     (print_type i s.sym_name) lhs;
-  let rhs = Basics.term_of_rhs r in
+  let rhs = LibTerm.term_of_rhs r in
   Format.fprintf oc "<TLrhs>@.%a</TLrhs>@.</typeLevelRule>@."
     (print_type i s.sym_name) rhs
 
 (** [get_vars s r] returns the list of variables used in the rule [r],
     in the form of a pair containing the name of the variable and its type,
     inferred by the solver. *)
-let get_vars : sym -> rule -> (string * Terms.term) list = fun s r ->
+let get_vars : sym -> rule -> (string * Term.term) list = fun s r ->
   let rule_ctx : tvar option array = Array.make (Array.length r.vars) None in
   let var_list : tvar list ref = ref [] in
   let rec subst_patt v t =