+++ tests/ok_ko.ml
@@ -1,7 +1,6 @@
 (** Compile files in "OK" and "KO". *)
-open Core
 
-let compile = Compile.Pure.compile_file
+let compile = Handle.Compile.Pure.compile_file
 
 let test_ok f () =
   (* Simply assert that there is no exception raised. *)
@@ -13,7 +12,7 @@ let test_ko f () =
   Alcotest.(check bool) f r false
 
 let _ =
-  Files.set_lib_root None;
+  Common.Module.set_lib_root None;
   let open Alcotest in
   let files = Sys.readdir "OK" |> Array.map (fun f -> "OK/" ^ f)
 (* TODO put back OK/unif_hint.lp when it is fixed *)