+++ CHANGES.md
@@ -1,5 +1,27 @@
 ### Unreleased
 
+#### File renamings and source code segmentation (2021-02-??)
+
+- Files renamings:
+  * `terms` -> `term`
+  * `basics` -> `libTerm`
+  * `tactics` -> `tactic`
+  * `queries` -> `query`
+  * `stubs` -> `realpath`
+  * `files` -> `module`
+  * `handle` -> `command`
+
+- `Core` library divided into the following sub-libraries: 
+  * `Common` that contains shared basic files
+    (`pos`, `console`, `module` and `package`)
+  * `Parsing` that contains everything related to parsing
+    (`syntax`, `pretty`, lexers and parser)
+  * `Handle` that uses `Core` to type check commands and modules 
+    (contains `query`, `tactic`, `command`, `compile`, `inductive`, `rewrite`,
+    `proof` and `why3_tactic`)
+  * `Tool` that provides miscellaneous tools that use `Core`
+    (`external`, `hrs`, `xtc`, `tree_graphviz`, `sr`)
+
 #### Add parameters to inductive definitions (2021-02-02)
 
 #### Parser (2021-01-30)